Resolved global variable symbol name conflicts that existed between EPANET (vars.h) and MSX (globals.h).  This allows the EPANET and MSX code to be compiled together without first building and linking to separate DLLs.  This in turn allows for building and proper execution of MSX on Linux using shared object libs.  It is verified that the code produces identical results to those prior to this change.  No substantive algorithm changes were made - i.e. simply resolving the global symbol conflicts.
Specifically, ALL global variable symbols defined in globals.h were
made unique by prepending 'MSX' to their previous names.  This was
also done for all data types defined in typesmsx.h but used to
declare global symbols in globals.h.

It is proposed that any future modifications to MSX also use the
rule that any new global symbols should begin with 'MSX' so as to
practically distinguish them from any other EPANET globals.

The new global variable symbols were substituted in the following
files:

    MSX/trunk/src/rk5.c
    MSX/trunk/src/reportmsx.c
    MSX/trunk/src/epanetmsx.c
    MSX/trunk/src/project.c
    MSX/trunk/src/chemistry.c
    MSX/trunk/src/msxfile.c
    MSX/trunk/src/newton.c
    MSX/trunk/src/toolkitmsx.c
    MSX/trunk/src/qualitymsx.c
    MSX/trunk/src/ros2.c
    MSX/trunk/src/inputmsx.c
    MSX/trunk/src/outputmsx.c
    MSX/trunk/src/mathexpr.c

With one or two exceptions (caught by the compiler) all global
substitutions were made using 'sed -f modifyglobals.sed <file.c>'
where <file.c> is one of the above MSX *.c files and
modifyglobals.sed is a text file containing the following sed
commands:

s!\([^a-zA-Z0-9\"]\)HydFile\([^a-zA-Z0-9\"]\)!\1MSXHydFile\2!g
s!\([^a-zA-Z0-9\"]\)InpFile\([^a-zA-Z0-9\"]\)!\1MSXInpFile\2!g
s!\([^a-zA-Z0-9\"]\)OutFile\([^a-zA-Z0-9\"]\)!\1MSXOutFile\2!g
s!\([^a-zA-Z0-9\"]\)RptFile\([^a-zA-Z0-9\"]\)!\1MSXRptFile\2!g
s!\([^a-zA-Z0-9\"]\)Title\([^a-zA-Z0-9\"]\)!\1MSXTitle\2!g
s!\([^a-zA-Z0-9\"]\)Msg\([^a-zA-Z0-9\"]\)!\1MSXMsg\2!g
s!\([^a-zA-Z0-9\"]\)Nobjects\([^a-zA-Z0-9\"]\)!\1MSXNobjects\2!g
s!\([^a-zA-Z0-9\"]\)Unitsflag\([^a-zA-Z0-9\"]\)!\1MSXUnitsflag\2!g
s!\([^a-zA-Z0-9\"]\)Flowflag\([^a-zA-Z0-9\"]\)!\1MSXFlowflag\2!g
s!\([^a-zA-Z0-9\"]\)Saveflag\([^a-zA-Z0-9\"]\)!\1MSXSaveflag\2!g
s!\([^a-zA-Z0-9\"]\)Rptflag\([^a-zA-Z0-9\"]\)!\1MSXRptflag\2!g
s!\([^a-zA-Z0-9\"]\)AreaUnits\([^a-zA-Z0-9\"]\)!\1MSXAreaUnits\2!g
s!\([^a-zA-Z0-9\"]\)RateUnits\([^a-zA-Z0-9\"]\)!\1MSXRateUnits\2!g
s!\([^a-zA-Z0-9\"]\)Solver\([^a-zA-Z0-9\"]\)!\1MSXSolver\2!g
s!\([^a-zA-Z0-9\"]\)PageSize\([^a-zA-Z0-9\"]\)!\1MSXPageSize\2!g
s!\([^a-zA-Z0-9\"]\)Nperiods\([^a-zA-Z0-9\"]\)!\1MSXNperiods\2!g
s!\([^a-zA-Z0-9\"]\)ErrCode\([^a-zA-Z0-9\"]\)!\1MSXErrCode\2!g
s!\([^a-zA-Z0-9\"]\)ProjectOpened\([^a-zA-Z0-9\"]\)!\1MSXProjectOpened\2!g
s!\([^a-zA-Z0-9\"]\)QualityOpened\([^a-zA-Z0-9\"]\)!\1MSXQualityOpened\2!g
s!\([^a-zA-Z0-9\"]\)HydOffset\([^a-zA-Z0-9\"]\)!\1MSXHydOffset\2!g
s!\([^a-zA-Z0-9\"]\)Qstep\([^a-zA-Z0-9\"]\)!\1MSXQstep\2!g
s!\([^a-zA-Z0-9\"]\)Pstep\([^a-zA-Z0-9\"]\)!\1MSXPstep\2!g
s!\([^a-zA-Z0-9\"]\)Pstart\([^a-zA-Z0-9\"]\)!\1MSXPstart\2!g
s!\([^a-zA-Z0-9\"]\)Rstep\([^a-zA-Z0-9\"]\)!\1MSXRstep\2!g
s!\([^a-zA-Z0-9\"]\)Rstart\([^a-zA-Z0-9\"]\)!\1MSXRstart\2!g
s!\([^a-zA-Z0-9\"]\)Rtime\([^a-zA-Z0-9\"]\)!\1MSXRtime\2!g
s!\([^a-zA-Z0-9\"]\)Htime\([^a-zA-Z0-9\"]\)!\1MSXHtime\2!g
s!\([^a-zA-Z0-9\"]\)Qtime\([^a-zA-Z0-9\"]\)!\1MSXQtime\2!g
s!\([^a-zA-Z0-9\"]\)Dur\([^a-zA-Z0-9\"]\)!\1MSXDur\2!g
s!\([^a-zA-Z0-9\"]\)D\([^a-zA-Z0-9\"]\)!\1MSXD\2!g
s!\([^a-zA-Z0-9\"]\)H\([^a-zA-Z0-9\"]\)!\1MSXH\2!g
s!\([^a-zA-Z0-9\"]\)Q\([^a-zA-Z0-9\"]\)!\1MSXQ\2!g
s!\([^a-zA-Z0-9\"]\)Ucf\([^a-zA-Z0-9\"]\)!\1MSXUcf\2!g
s!\([^a-zA-Z0-9\"]\)DefRtol\([^a-zA-Z0-9\"]\)!\1MSXDefRtol\2!g
s!\([^a-zA-Z0-9\"]\)DefAtol\([^a-zA-Z0-9\"]\)!\1MSXDefAtol\2!g
s!\([^a-zA-Z0-9\"]\)Specie\([^a-zA-Z0-9\"]\)!\1MSXSpecie\2!g
s!\([^a-zA-Z0-9\"]\)Param\([^a-zA-Z0-9\"]\)!\1MSXParam\2!g
s!\([^a-zA-Z0-9\"]\)Const\([^a-zA-Z0-9\"]\)!\1MSXConst\2!g
s!\([^a-zA-Z0-9\"]\)Term\([^a-zA-Z0-9\"]\)!\1MSXTerm\2!g
s!\([^a-zA-Z0-9\"]\)Node\([^a-zA-Z0-9\"]\)!\1MSXNode\2!g
s!\([^a-zA-Z0-9\"]\)Link\([^a-zA-Z0-9\"]\)!\1MSXLink\2!g
s!\([^a-zA-Z0-9\"]\)Tank\([^a-zA-Z0-9\"]\)!\1MSXTank\2!g
s!\([^a-zA-Z0-9\"]\)Pattern\([^a-zA-Z0-9\"]\)!\1MSXPattern\2!g



git-svn-id: https://epanet.svn.sourceforge.net/svnroot/epanet/BASE/trunk@37 c320cabd-cc23-0410-96d8-e60fbf53ed7f